BUNDLE 1. more long int move costs conversions + Passenger_Move_Cost bug fix
This commit also requires the next. This one tries to mostly separate the continued conversions from int to long int, from the Passenger_Move_Cost bug fix, but there is overlap in the same files as they were fixed simultaneously when speculating they had same cause.

+++ freeciv/freeciv/ai/default/aiair.c
@@ -332,7 +332,7 @@ void dai_manage_airunit(struct ai_type *ait, struct player *pplayer,
 {
   struct tile *dst_tile = unit_tile(punit);
   /* Loop prevention */
-  int moves = punit->moves_left;
+  long moves = punit->moves_left;
   int id = punit->id;
   struct pf_parameter parameter;
   struct pf_map *pfm;
